Organize directory structure in views, models and components
