+++ lib/views/note_view.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:note_taker/main.dart';
 import 'package:note_taker/models/note_model.dart';
+
+import 'home_page.dart';
 
 class NoteView extends StatefulWidget {
   NoteView({Key key, this.notes, this.index}) : super(key: key);
